chore(monorepo): update configs, documentation, windows app and conductor

+++ windows_app/lib/shared/widgets/optimized_list_view.dart
@@ -17,7 +17,7 @@ class OptimizedListView<T> extends StatelessWidget {
   final bool shrinkWrap;
   final Axis scrollDirection;
   final double? itemExtent;
-  final int? cacheExtent;
+  final double? cacheExtent;
   final bool addAutomaticKeepAlives;
   final bool addRepaintBoundaries;
   final bool addSemanticIndexes;
